Refactor: 환불 상태 공통화와 booking helper 트랜잭션 경계 강화
