+++ CLAUDE.md
@@ -57,6 +57,6 @@
 
 ---
 
-## 역할 분담
-- 이 에이전트는 **구현(코드 작성)에만 집중**.
-- 테스트·문서 동기화는 별도 에이전트 담당 — 직접 수행하지 않는다.
+## 문서 동기화
+- 소스 수정 시 관련 문서(`docs/PRD/`, `docs/ADR/`, `simple-idea.md` 등)도 함께 갱신한다.
+- 스킬의 **doc sync checklist**를 기준으로 갱신 대상을 판단한다.
